Add U21 print option (#12)
A `u21` is a weirdly specific thing to use as a number, I thought it
would be nice if it printed as a character by default. I think this 
is correct for pretty also. If we say "u21s are characters" then 
silently switching to decimal printing when they don't fit is a weird
thing to do.

* Add u21_is_codepoint option, default true
* Test over-sized u21. Turns out Zig doesn't panic (good) and uses 
  the replacement character.
* Update comment and README